+++ packages/sdk-next/src/api/index.ts
@@ -1,2 +1,2 @@
-export { Router } from './Router';
-export { TradeRouter } from './TradeRouter';
+export * from './client';
+export * from './Papi';