+++ packages/sdk-next/src/index.ts
@@ -2,8 +2,8 @@ export * as api from './api';
 export * as client from './client';
 export * as const from './consts';
 export * as error from './errors';
-export * as papi from './provider';
 export * as pool from './pool';
+export * as sor from './sor';
 
 export * from './types';
 export * from './utils';