+++ packages/sdk-next/src/utils/erc20.ts
@@ -1,0 +1,7 @@
+import { Buffer } from 'buffer';
+
+export function convertToId(xc: string): number {
+  const addressBuffer = Buffer.from(xc, 'hex');
+  const assetIdBuffer = addressBuffer.subarray(16);
+  return assetIdBuffer.readUIntBE(0, assetIdBuffer.length);
+}