+++ packages/sdk-next/test/script/PapiExecutor.ts
@@ -3,7 +3,7 @@ import { hydration } from '@polkadot-api/descriptors';
 
 import { ApiUrl } from './types';
 
-import { papi, json } from '../../src';
+import { api, json } from '../../src';
 
 export abstract class PapiExecutor {
   protected readonly apiUrl: ApiUrl;
@@ -17,9 +17,9 @@ export abstract class PapiExecutor {
   }
 
   async run() {
-    const client = await papi.getWs(this.apiUrl);
-    const api = client.getTypedApi(hydration);
-    const { spec_name, spec_version } = await api.constants.System.Version();
+    const client = await api.getWs(this.apiUrl);
+    const papi = client.getTypedApi(hydration);
+    const { spec_name, spec_version } = await papi.constants.System.Version();
     console.log(`Runtime ready ${spec_name}/${spec_version}`);
     console.log('Running script...');
     console.log(this.desc);
@@ -27,6 +27,11 @@ export abstract class PapiExecutor {
 
     this.script(client)
       .then((output: any) => {
+        if (typeof output === 'function') {
+          setTimeout(output, 1000);
+          return;
+        }
+
         if (this.pretty) {
           console.log(
             output ? JSON.stringify(output, json.jsonFormatter, 2) : ''
@@ -34,7 +39,7 @@ export abstract class PapiExecutor {
         } else {
           console.log(output);
         }
-        return null;
+        client.destroy();
       })
       .catch((e) => {
         console.log(e);
@@ -42,9 +47,17 @@ export abstract class PapiExecutor {
       })
       .finally(() => {
         console.timeEnd('Execution time:');
-        client.destroy();
       });
   }
 
   abstract script(client: PolkadotClient): Promise<any>;
+
+  logTime() {
+    const time = [
+      '-----',
+      new Date().toISOString().replace('T', ' ').replace('Z', ''),
+      '-----',
+    ].join('');
+    console.log(time);
+  }
 }