+++ packages/xcm-cfg/src/builders/extrinsics/xcm/polkadotXcm.utils.ts
@@ -1,4 +1,9 @@
-import { multiloc, Parachain, TxWeight } from '@galacticcouncil/xcm-core';
+import {
+  ChainEcosystem,
+  multiloc,
+  Parachain,
+  TxWeight,
+} from '@galacticcouncil/xcm-core';
 
 import { getX1Junction } from './utils';
 import { XcmTransferType, XcmVersion } from './types';
@@ -12,7 +17,45 @@ const ETHEREUM_BRIDGE_LOCATION = {
 };
 const HUB_PARACHAIN_ID = 1000;
 
-export const toDest = (version: XcmVersion, destination: Parachain) => {
+const BRIDGE_CONSENSUS = [ChainEcosystem.Polkadot, ChainEcosystem.Kusama];
+
+const isBridgeHubTransfer = (source: Parachain, destination: Parachain) => {
+  const isUnknownConsensus = !source.ecosystem || !destination.ecosystem;
+  if (isUnknownConsensus) {
+    return false;
+  }
+
+  return (
+    BRIDGE_CONSENSUS.includes(source.ecosystem) &&
+    BRIDGE_CONSENSUS.includes(destination.ecosystem) &&
+    source.ecosystem !== destination.ecosystem
+  );
+};
+
+export const toDest = (
+  version: XcmVersion,
+  source: Parachain,
+  destination: Parachain
+) => {
+  if (isBridgeHubTransfer(source, destination)) {
+    return {
+      [version]: {
+        parents: 2,
+        interior: {
+          X2: [
+            {
+              GlobalConsensus: destination.ecosystem,
+            },
+            {
+              Parachain: destination.parachainId,
+            },
+          ],
+        },
+      },
+    };
+  }
+
+  // Relay transfer
   if (destination.parachainId === 0) {
     return {
       [version]: {
@@ -22,17 +65,17 @@ export const toDest = (version: XcmVersion, destination: Parachain) => {
     };
   }
 
-  const toParachain = {
-    Parachain:
-      destination.key === 'ethereum'
-        ? HUB_PARACHAIN_ID
-        : destination.parachainId,
-  };
+  // Ethereum snowbridge transfers are routed via hub
+  const parachain =
+    destination.key === 'ethereum' ? HUB_PARACHAIN_ID : destination.parachainId;
+
   return {
     [version]: {
       parents: 1,
       interior: {
-        X1: getX1Junction(version, toParachain),
+        X1: getX1Junction(version, {
+          Parachain: parachain,
+        }),
       },
     },
   };