+++ packages/xcm-cfg/src/configs/kusama/basilisk/index.ts
@@ -13,6 +13,7 @@ import {
 import { ExtrinsicBuilder } from '../../../builders';
 
 import { balance, fee } from './configs';
+import { toTransferTemplate } from './templates';
 
 const toAssetHub: AssetRoute[] = [
   new AssetRoute({
@@ -28,7 +29,7 @@ const toAssetHub: AssetRoute[] = [
       chain: kusamaAssetHub,
       asset: usdt,
       fee: {
-        amount: 0.001183,
+        amount: 0.0012,
         asset: usdt,
       },
     },
@@ -36,124 +37,14 @@ const toAssetHub: AssetRoute[] = [
   }),
 ];
 
-const toKarura: AssetRoute[] = [
-  new AssetRoute({
-    source: {
-      asset: bsx,
-      balance: balance(),
-      fee: fee(),
-      destinationFee: {
-        balance: balance(),
-      },
-    },
-    destination: {
-      chain: karura,
-      asset: bsx,
-      fee: {
-        amount: 0.09324,
-        asset: bsx,
-      },
-    },
-    extrinsic: ExtrinsicBuilder().xTokens().transfer(),
-  }),
-];
-
-const toKusama: AssetRoute[] = [
-  new AssetRoute({
-    source: {
-      asset: ksm,
-      balance: balance(),
-      fee: fee(),
-      destinationFee: {
-        balance: balance(),
-      },
-    },
-    destination: {
-      chain: kusama,
-      asset: ksm,
-      fee: {
-        amount: 0.00010457164,
-        asset: ksm,
-      },
-    },
-    extrinsic: ExtrinsicBuilder().xTokens().transfer(),
-  }),
-];
-
-const toIntegritee: AssetRoute[] = [
-  new AssetRoute({
-    source: {
-      asset: teer,
-      balance: balance(),
-      fee: fee(),
-      destinationFee: {
-        balance: balance(),
-      },
-    },
-    destination: {
-      chain: integritee,
-      asset: teer,
-      fee: {
-        amount: 0.000004,
-        asset: teer,
-      },
-    },
-    extrinsic: ExtrinsicBuilder().xTokens().transfer(),
-  }),
-];
-
-const toRobonomics: AssetRoute[] = [
-  new AssetRoute({
-    source: {
-      asset: xrt,
-      balance: balance(),
-      fee: fee(),
-      destinationFee: {
-        balance: balance(),
-      },
-    },
-    destination: {
-      chain: robonomics,
-      asset: xrt,
-      fee: {
-        amount: 0.000004632,
-        asset: xrt,
-      },
-    },
-    extrinsic: ExtrinsicBuilder().xTokens().transfer(),
-  }),
-];
-
-const toTinkernet: AssetRoute[] = [
-  new AssetRoute({
-    source: {
-      asset: tnkr,
-      balance: balance(),
-      fee: fee(),
-      destinationFee: {
-        balance: balance(),
-      },
-    },
-    destination: {
-      chain: tinkernet,
-      asset: tnkr,
-      fee: {
-        amount: 0.00927020324,
-        asset: tnkr,
-      },
-    },
-    extrinsic: ExtrinsicBuilder().xTokens().transfer(),
-  }),
-];
-
 export const basiliskConfig = new ChainRoutes({
   chain: basilisk,
   routes: [
     ...toAssetHub,
-    ...toKarura,
-    ...toKusama,
-    ...toIntegritee,
-    ...toRobonomics,
-    ...toTinkernet,
+    toTransferTemplate(bsx, karura, 0.0933),
+    toTransferTemplate(ksm, kusama, 0.00012),
+    toTransferTemplate(teer, integritee, 0.000004),
+    toTransferTemplate(xrt, robonomics, 0.00000464),
+    toTransferTemplate(tnkr, tinkernet, 0.0095),
   ],
 });