+++ packages/xcm-sdk/src/platforms/substrate/utils/dryRun.ts
@@ -1,3 +1,7 @@
+import { ApiPromise } from '@polkadot/api';
+
+import { BN } from '@polkadot/util';
+
 export const getDeliveryFeeFromDryRun = (ok: any): bigint => {
   const deliveryFees: bigint[] = [];
 
@@ -14,4 +18,20 @@ export const getDeliveryFeeFromDryRun = (ok: any): bigint => {
     }
   }
   return deliveryFees.reduce((acc, df) => acc + df, 0n);
+};
+
+export const getErrorFromDryRun = (api: ApiPromise, ok: any): string => {
+  const err = ok.executionResult.Err.error;
+  if (err.Module) {
+    const { error, index } = err.Module;
+
+    const errorIndex = Number(error.slice(0, 4));
+    const decoded = api.registry.findMetaError({
+      error: new BN(errorIndex),
+      index: new BN(index),
+    });
+    return `${decoded.section}.${decoded.method}: ${decoded.docs.join(' ')}`;
+  } else {
+    return JSON.stringify(err);
+  }
 };